update(ot):type提取到types.ts文件
